Fix warnings, improve code style

+++ src/QtWebApp/httpserver/staticfilecontroller.cpp
@@ -33,7 +33,7 @@ StaticFileController::StaticFileController(QSettings* settings, QObject* parent)
     maxCachedFileSize=settings->value("maxCachedFileSize","65536").toInt();
     cache.setMaxCost(settings->value("cacheSize","1000000").toInt());
     cacheTimeout=settings->value("cacheTime","60000").toInt();
-    qDebug("StaticFileController: cache timeout=%i, size=%i",cacheTimeout,cache.maxCost());
+    qDebug("StaticFileController: cache timeout=%i, size=%lld",cacheTimeout,static_cast<long long>(cache.maxCost()));
 }
